AD_1296: Removing previous change and adding support for parameter targeting highlighting. By default now all query hexes always highlights the selected hexs by default. Only creature movement sets it to false, so you get the dark path. Selecting a choice path, like Grumble's Boom Box will not highlight the entire path as specified in 1296, and so should any other ability using the same targeting method queryChoice.
updateDisplay was being abused, and was being called everywhere. It shouldn't be. Hexegons should be controlling cleaning the display. Now whenever a hex needs to be changed, cleanHex should be called instead to clean only that hex and then apply the changes directly to it, instead of clearing the entire screen.